refactor: Break utility to mass-insert GTFS test data to its own module
